requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langchain-classic>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=4.0.0
python-dotenv>=1.0.0
langchain-huggingface